Java 1.8 compatibility changes, gitignore oversights, and a javadocs bugfix
Made this project Java 1.8 compatible, and removed the module-info file (saved a copy first), with the Maven compatibility also bumped back for Java 1.8 issues with their latest update. Added Eclipse project oriented files, which are generic and system-agnostic. Fixed a missing include in FunctionType.java that caused an error in building the javadocs.